rvv: fix vssra.vi e64 corner case

+++ riscv/insns/vssra_vi.h
@@ -3,7 +3,7 @@ VRM xrm = P.VU.get_vround_mode();
 VI_VI_LOOP
 ({
   int sh = simm5 & (sew - 1) & 0x1f;
-  int64_t val = vs2;
+  int128_t val = vs2;
 
   INT_ROUNDING(val, xrm, sh);
   vd = val >> sh;
